% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandpart.R
\name{grandpart}
\alias{grandpart}
\title{Accessory Function for Cross-validation}
\usage{
grandpart(dataset, k)
}
\arguments{
\item{dataset}{a dataset for partitioning.}

\item{k}{the number of partitions; the cross-validation fold number.}
}
\value{
Returns a list containing the partitioned datasets.
}
\description{
Randomly partition a dataset into k approximately equally sized partitions.
}
\note{
This function is not designed to be called directly,
      but acts within \code{kcrossval}
}

