% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randboot.R
\name{randboot}
\alias{randboot}
\title{Bootstrap Resampling}
\usage{
randboot(dataset, n, replace = TRUE)
}
\arguments{
\item{dataset}{a p x m data matrix.}

\item{n}{the number of rows of observations to be included in the new dataset.}

\item{replace}{logical. If replace is TRUE sampling will be with replacement, thus
returning a bootstrap replicate.}
}
\value{
The function returns a new n x m matrix. The default is a bootstrap
        replicate of the orginal data.
}
\description{
Generate a dataset from a given dataset using sampling with or without replacement.
}
\details{
This function is a simple shortcut for generating a bootstrap sample. It was
originally designed for use within the comparison framework, but can be used as a
stand-alone function.
}

