% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randpart.R
\name{randpart}
\alias{randpart}
\title{Accessory Function for Sample Splitting}
\usage{
randpart(dataset, fract)
}
\arguments{
\item{dataset}{a dataset for splitting.}

\item{fract}{the proportion of observations to be designated to the training set.}
}
\value{
Returns a list containing the training set and test set.
}
\description{
Randomly partition a dataset into a training set and a test set.
}
\note{
This function is not designed to be called directly,
      but acts within \code{splitval}
}

